sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vicinal"
version = "0.1.0"
description = "Runtime AI guardrail SDK — FAISS-based zero-resource + OCR/RAG context-driven prompt safety evaluation."
authors = [{ name = "Vicinal Contributors" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai", "safety", "guardrails", "llm", "prompt-injection", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies — always installed
dependencies = [
    "sentence-transformers>=2.7,<3.0",
    "faiss-cpu>=1.8,<2.0",
    "numpy>=1.26,<3.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
# pip install vicinal[ocr]
ocr = [
    "easyocr>=1.7",
    "pymupdf>=1.24",   # PDF rasterisation
]

# pip install vicinal[rag]
rag = [
    "rank-bm25>=0.2",
]

# pip install vicinal[full]  — everything
full = [
    "easyocr>=1.7",
    "pymupdf>=1.24",
    "rank-bm25>=0.2",
]

# pip install vicinal[webapp]  — demo web application
webapp = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]

# pip install vicinal[research]  — baseline comparison framework
research = [
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "rank-bm25>=0.2",
]

# pip install vicinal[dev]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["vicinal*"]

[tool.setuptools.package-data]
vicinal = ["py.typed"]

# -----------------------------------------------------------------------
# Tooling
# -----------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
